+++ lib/state_machine.dart
@@ -1,9 +1,5 @@
 library state_machine;
 
-import 'dart:collection';
-
-import 'package:binary_expression_tree/binary_expression_tree.dart';
-
 part 'src/state_machine.dart';
 part 'src/state.dart';
 part 'src/transition.dart';