+++ test/state_test.dart
@@ -3,23 +3,41 @@ import 'package:test/test.dart';
 
 void main() {
   group('State', () {
-    test('can be created', () {
-      StateMachine machine = StateMachine('test');
-      State state = machine.newState('test');
-      expect(state.name, equals('test'));
-      expect(machine.states.contains(state), isTrue);
-      expect(state.transitions, isEmpty);
-    });
-    test('can be added to a state machine', () {
-      StateMachine machine = StateMachine('test');
-      State state = machine.newState('test');
-      expect(machine.states, [state]);
+    late StateMachine machine;
+    late State state;
+
+    setUp(() {
+      machine = StateMachine('test');
+      state = machine.newState('test');
     });
 
-    test('can be added to a state machine only once', () {
-      StateMachine machine = StateMachine('test');
-      State state = machine.newState('test');
-      expect(() => machine.newState('test'), throwsException);
+    test('Expect state name equals name', () {
+      expect(state.name, equals('test'));
+    });
+
+    test('Expect machine contains state', () {
+      expect(machine.states, contains(state));
+    });
+
+    test('Expect transitions is empty', () {
+      expect(state.transitions, isEmpty);
+    });
+
+    test('Expect transitions contains transition', () {
+      State state2 = machine.newState('test2');
+      machine.newTransition('test', {state}, state2);
+      expect(state.transitions, hasLength(1));
+      expect(state.transitions, equals(machine.transitions[state]));
+    });
+
+    test('Expect machine isActive equals false', () {
+      State state2 = machine.newState('test2');
+      expect(state2.isActive, isFalse);
+    });
+
+    test('Expect machine isActive equals true', () {
+      machine.start(state);
+      expect(state.isActive, isTrue);
     });
   });
 }