+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- Added tests for state machine, state and transition
+- Decoupled library
+
 ## 1.0.0
 
 - Initial version.